+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@
 - added animated interaction support to keyholes and puzzle slots (#4471)
 - added an option to control snap interactions for targets such as pickups and switches (Gameplay → Controls → Snap interactions)
 - changed vehicles to allow defining in Lua whether or not they can activate heavy triggers; refer to migration notes
+- change the Skidoo and Quad Bike to allow defining in Lua whether or not they should collide with static meshes (#4285)
 - changed Assault Course Lua record access from `trx.assault_stats` to `trx.assault.stats`
 - changed hard-coded music tracks from Snowmobiles, mine carts, and RIBs to be configurable via Lua
 - changed hard-coded fish and piranha setup to be configurable via Lua